+++ backend/Carma.Application/DTOs/User/UserProfileDto.cs
@@ -3,6 +3,6 @@ namespace Carma.Application.DTOs.User;
 public record UserProfileDto(
     string UserName,
     string? ImageUrl,
-    int Karma,
+    double Karma,
     int RidesCount
     );
