+++ backend/Carma.Application/DTOs/User/UserSummaryDto.cs
@@ -3,5 +3,5 @@ namespace Carma.Application.DTOs.User;
 public record UserSummaryDto(
     string UserName,
     string? ImageUrl,
-    int Karma
+    double Karma
     );
